girly-to-do-list
This to-do list website features a soft and elegant design with pastel colors, playful fonts, and charming icons. I created the custom logo as well as the checked and unchecked icons to give the site a personal touch. It allows you to easily add, manage, and check off your daily tasks while enjoying a visually delightful experience. Perfect for anyone who loves blending productivity with a sprinkle of girly aesthetics.